[MC-15] [api] refactor: change import path name and update adapter logic to align with the refactored user domain logic.
- Rename utils directory to commons.
- Change directory name from auth to user in the user domain area.
[related only to adapter objects below]
- Change from AuthService to UserMetadataRepository code
- Update SQL queries in SQL builder to align with the changes in UserMetadata.
