+++ apps/api/src/numerical-guidance/infrastructure/adapter/persistence/history-indicator/entity/history-indicator.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
 import { HistoryIndicatorValueEntity } from '../../history-indicator-value/entity/history-indicator-value.entity';
-import { IndicatorType } from '../../../../../../utils/type/type-definition';
+import { IndicatorType } from '../../../../../../commons/type/type-definition';
 
 @Entity({ name: 'historyIndicator' })
 export class HistoryIndicatorEntity {
