+++ apps/api/src/numerical-guidance/infrastructure/adapter/persistence/indicator/entity/bonds.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity } from 'typeorm';
 import { BaseIndicatorEntity } from './base-indicator.entity';
-import { IndicatorType } from '../../../../../../utils/type/type-definition';
+import { IndicatorType } from '../../../../../../commons/type/type-definition';
 
 @Entity({ name: 'bonds' })
 export class BondsEntity extends BaseIndicatorEntity {
