+++ apps/api/src/numerical-guidance/infrastructure/adapter/persistence/indicator/entity/forex-pair.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity } from 'typeorm';
 import { BaseIndicatorEntity } from './base-indicator.entity';
-import { IndicatorType } from '../../../../../../utils/type/type-definition';
+import { IndicatorType } from '../../../../../../commons/type/type-definition';
 
 @Entity({ name: 'forex_pairs' })
 export class ForexPairEntity extends BaseIndicatorEntity {
