+++ apps/api/src/numerical-guidance/infrastructure/adapter/persistence/indicator/indicator.persistent.adapter.ts
@@ -11,7 +11,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { LoadIndicatorPort } from '../../../../application/port/persistence/indicator/load-indicator.port';
 import { TypeORMError } from 'typeorm/error/TypeORMError';
 import { LoadIndicatorListPort } from '../../../../application/port/persistence/indicator/load-indicator-list.port';
-import { IndicatorDtoType, IndicatorEntityType, IndicatorType } from '../../../../../utils/type/type-definition';
+import { IndicatorDtoType, IndicatorEntityType, IndicatorType } from '../../../../../commons/type/type-definition';
 import { BondsEntity } from './entity/bonds.entity';
 import { CryptoCurrenciesEntity } from './entity/crypto-currencies.entity';
 import { ETFEntity } from './entity/etf.entity';
@@ -19,8 +19,8 @@ import { ForexPairEntity } from './entity/forex-pair.entity';
 import { FundEntity } from './entity/fund.entity';
 import { IndicesEntity } from './entity/indices.entity';
 import { StockEntity } from './entity/stock.entity';
-import { CursorPageMetaDto } from '../../../../../utils/pagination/cursor-page.meta.dto';
-import { CursorPageDto } from '../../../../../utils/pagination/cursor-page.dto';
+import { CursorPageMetaDto } from '../../../../../commons/pagination/cursor-page.meta.dto';
+import { CursorPageDto } from '../../../../../commons/pagination/cursor-page.dto';
 import { SaveIndicatorListPort } from '../../../../application/port/external/twelve/save-indicator-list.port';
 import { Propagation, Transactional } from 'typeorm-transactional';
 import { TwelveApiManager } from '../../twelve/util/twelve-api.manager';
